Modify pushing of new studies to global kbase
Modified the GlobalKnowledgebase method of receiving a new study, such that the publication is stochastic (Bernoulli trial), and depends on the parameter of the model - `p_study_published`.
To make this possible, GlobalKnowledgebase class also now requires a reference to the model (to be able to access the model parameter governing stochasticity).